requirements.txt
# File: requirements.txt

google-generativeai
ebooklib
PyMuPDF
beautifulsoup4
python-dotenv